refactor: Extract getMimetypeFromUrl into files utils file
